% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdispersion.R
\name{ggdispersion}
\alias{ggdispersion}
\title{Plot VAR Residuals Dispersion}
\usage{
ggdispersion(
  x,
  ncol = 1,
  alpha = 0.5,
  palette = c(points = "black", xaxis = "black")
)
}
\arguments{
\item{x}{A "varest" object to get residuals and fitted values from.}

\item{ncol}{A interger. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{alpha}{A double. The alpha aesthetic for the points, passed to \link[ggplot2]{geom_point}.}

\item{palette}{A vector of colors. Each for a different ggplot2 element, as denoted by the (optional) names.}
}
\value{
Returns an object of class \code{ggplot}.
}
\description{
Plots a scatterplot of the residuals versus fitted values of a VAR model, using ggplot2.
}
\examples{
x <- VAR(EuStockMarkets)
ggdispersion(x)
}
\seealso{
For info on \code{alpha} and \code{colors}: \url{\code{vignette("ggplot2-specs")}}.
}
