% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_acf.R
\name{ggvar_acf}
\alias{ggvar_acf}
\alias{ggvar_ccf_grid}
\alias{ggvar_ccf_ind}
\alias{ggvar_ccf_wrap}
\title{Plot autocorrelation (and similars) of dataset}
\usage{
ggvar_acf(
  x,
  series = NULL,
  type = "correlation",
  lag.max = NULL,
  ci = 0.95,
  ...,
  geom = "segment",
  args_geom = list(),
  args_ribbon = list(linetype = 2, color = "blue", fill = NA),
  args_hline = list(),
  args_facet = list()
)

ggvar_ccf_grid(
  x,
  series = NULL,
  type = "correlation",
  lag.max = NULL,
  ci = 0.95,
  ...,
  geom = "segment",
  facet = "ggplot",
  args_geom = list(),
  args_ribbon = list(linetype = 2, color = "blue", fill = NA),
  args_hline = list(),
  args_facet = list()
)

ggvar_ccf_ind(x, serie_x, serie_y, ...)

ggvar_ccf_wrap(x, series, ...)
}
\arguments{
\item{x}{A dataset (object coercible to data.frame) or a "varest" object to
get residuals from.}

\item{series}{A character vector with variables (column names) to consider. Defaults to all (\code{NULL}).}

\item{type}{The type of ACF to be computed, passed to \link[stats]{acf}. Can
be either "correlation", "covariance", or "partial".}

\item{lag.max}{The number of lags used to calculate the ACF, passed to
\link[stats]{acf}. defaults to \code{10 * log(nrow(x) / ncol(x), base = 10)}.}

\item{ci}{The level of confidence for the ACF confidence interval. Set to
\code{FALSE} to omit the \link[ggplot2]{geom_ribbon}.}

\item{...}{Additional arguments passed to \link[stats]{acf}.}

\item{geom}{The ggplot geom used to create the plot, 'segment' (the default) for \link[ggplot2]{geom_segment}, 'area' for \link[ggplot2]{geom_area}.}

\item{args_geom}{Arguments passed to the chosen \code{geom}.}

\item{args_ribbon}{Aditional arguments passed to \link[ggplot2]{geom_ribbon}.}

\item{args_hline}{Aditional arguments passed to \link[ggplot2]{geom_hline}.}

\item{args_facet}{Aditional arguments passed to \link[ggplot2]{facet_wrap}.}

\item{facet}{The facet "engine" to be used. 'ggplot2' for \link[ggplot2]{facet_grid}, 'ggh4x' for \link[ggh4x]{facet_grid2}}

\item{serie_x, serie_y}{Variables to chose for x and y axis in
\code{ggvar_ccf_ind}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
\code{ggvar_acf} plots the result of a \link[stats]{acf} call for every
 series, using \link[ggplot2]{facet_wrap}. \code{ggvar_ccf} plots all the
 cross correlations (and similars) between the series, using
 \link[ggplot2]{facet_grid}.
}
\examples{
ggvar_acf(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_ccf_grid(freeny[-2], args_facet = list(scales = "free_y"))
ggvar_acf(vars::VAR(freeny[-2]), args_facet = list(scales = "free_y"))

}
