% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_dispersion.R
\name{ggvar_dispersion}
\alias{ggvar_dispersion}
\title{Plot VAR Residuals Dispersion}
\usage{
ggvar_dispersion(
  x,
  series = NULL,
  args_point = list(),
  args_hline = list(),
  args_facet = list()
)
}
\arguments{
\item{x}{A "varest" object to get residuals and fitted values from.}

\item{series}{A character vector with variables (column names) to consider. Defaults to all (\code{NULL}).}

\item{args_point}{Aditional arguments passed to \link[ggplot2]{geom_point}.}

\item{args_hline}{Aditional arguments passed to \link[ggplot2]{geom_hline}.}

\item{args_facet}{Aditional arguments passed to \link[ggplot2]{geom_facet}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots a scatterplot of the residuals versus fitted values of a VAR model,
 using ggplot2.
}
\examples{
ggvar_dispersion(vars::VAR(freeny[-2]), args_facet = list(scales = "free_x"))

}
