% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_dist.R
\name{ggvar_dist}
\alias{ggvar_dist}
\title{Plot VAR Residuals Distribution}
\usage{
ggvar_dist(
  x,
  series = NULL,
  bins = 30,
  palette = c("grey", "black"),
  size = 1,
  ...
)
}
\arguments{
\item{x}{A "varest" object.}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{bins}{An interger. The number of histogram bins, passed to \link[ggplot2]{geom_histogram}.}

\item{palette}{A vector of colors (bins, normal curve). See \code{vignette("palettes")}.}

\item{size}{An interger. The size of the normal curve line, passed to \link[ggplot2]{geom_line}.}

\item{...}{Aditional arguments passed to \link[ggplot2]{geom_histogram}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the histogram of the residuals of a VAR model, overlayed with a normal curve.
}
\examples{
ggvar_dist(vars::VAR(EuStockMarkets))

}
