% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_fevd.R
\name{ggvar_fevd}
\alias{ggvar_fevd}
\title{Plot for Forecast Error Variance Decomposition of a VAR}
\usage{
ggvar_fevd(
  x,
  n.ahead = NULL,
  series = NULL,
  geom = "bar",
  colors = NULL,
  args_geom = list(),
  args_facet = list(),
  ...
)
}
\arguments{
\item{x}{A "varest" object to pass to \link[vars]{fevd}, or, directly, a
"varfevd" object.}

\item{n.ahead}{An integer. The size of the forecast horizon, passed to
\link[vars]{fevd}. Unused if `x` is a "varfevd" object.}

\item{series}{A character vector with variables (column names) to consider. Defaults to all (\code{NULL}).}

\item{geom}{The ggplot geom used to create the plot, 'segment' (the default) for \link[ggplot2]{geom_segment}, 'area' for \link[ggplot2]{geom_area}, 'line' for \link[ggplot2]{geom_line}.}

\item{colors}{A vector of colors for each variable. Passed to\link[ggplot2]{scale_color_manual}. See \code{vignette("colors")}.}

\item{args_geom}{Arguments passed to the chosen \code{geom}.}

\item{args_facet}{Aditional arguments passed to \link[ggplot2]{facet_wrap}.}

\item{...}{If \code{geom="line"}, additional arguments to
\link[ggplot2]{geom_point}}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the result of a \link[vars]{fevd} call.
}
\examples{
ggvar_fevd(vars::VAR(freeny[-2]), n.ahead = 10)

}
