% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_fit.R
\name{ggvar_fit}
\alias{ggvar_fit}
\title{Plot VAR Fitted Values}
\usage{
ggvar_fit(
  x,
  vars = NULL,
  date = 1:x$obs,
  compare = TRUE,
  scales = "fixed",
  ncol = 1,
  alpha = 0.5,
  palette = NULL,
  linetypes = c("solid", "dashed")
)
}
\arguments{
\item{x}{A "varest" object to get fitted values from.}

\item{vars}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{date}{A vector of labels to the x-axis, normally dates. Must have length equal to x$obs. Defaults to a vector of indexes.}

\item{compare}{Logical, should the true values be printed?}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{A interger. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{alpha}{A double. The alpha aesthetic for the points, passed to \link[ggplot2]{geom_point}.}

\item{palette}{A vector of colors. Each for a different variable.}

\item{linetypes}{A vecotr of linetypes (fitted, original), passed to \link[ggplot2]{scale_linetype_manual}.}
}
\value{
A ggplot.
}
\description{
Plots fitted values of a VAR model, versus the actual values.
}
