% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_select.R
\name{ggvar_select}
\alias{ggvar_select}
\title{Plot Information Criteria for VAR Lag Selection}
\usage{
ggvar_select(
  x,
  criteria = c("AIC", "HQ", "SC", "FPE"),
  trans = "none",
  args_line = list()
)
}
\arguments{
\item{x}{The result of a \link[vars]{VARselect} call.}

\item{criteria}{The criteria to be considered. Any of "AIC", "HQ", "SC",
and "FPE".}

\item{trans}{A transformation to apply to each criteria result (vector). Can
be a function, "none" (the default), or "index" to create index numbers.}

\item{args_line}{Aditional arguments passed to \link[ggplot2]{geom_line}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the result of a call to \link[vars]{VARselect}.
}
\examples{
ggvar_values(vars::VARselect(freeny[-2]), args_facet = list(scales = "free_y"))

}
