% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_acf.R
\name{ggvar_acf}
\alias{ggvar_acf}
\alias{ggvar_ccf}
\title{Plot Autocorrelation (and Similars) of Dataset}
\usage{
ggvar_acf(
  x,
  type = "correlation",
  lag.max = NULL,
  ci = 0.95,
  geom = "segment",
  facet = "ggplot",
  palette = c("black", "black", "blue", NA),
  scales = "fixed",
  ncol = 1,
  ...
)

ggvar_ccf(
  x,
  type = "correlation",
  lag.max = NULL,
  ci = 0.95,
  facet = "ggplot",
  geom = "segment",
  palette = c("black", "black", "blue", NA),
  scales = "fixed",
  independent = "y",
  ...
)
}
\arguments{
\item{x}{A dataset (object coercible to data.frame) or a "varest" object to get residuals from.}

\item{type}{The type of ACF to be computed, passed to \link[stats]{acf}. Can be either "correlation", "covariance", or "partial".}

\item{lag.max}{The number of lags used to calculate the ACF, passed to \link[stats]{acf}. defaults to \code{ceiling(10 * log(nrow(data) / ncol(data), base = 10))}.}

\item{ci}{The level of confidence for the ACF confidence interval. Set to \code{FALSE} to omit.}

\item{geom}{The ggplot geom used to create the plot, "segment" for \link[ggplot2]{geom_segment} (the default) or "area" for \link[ggplot2]{geom_area}.}

\item{facet}{The facet "engine" to be used. "ggplot2" for \link[ggplot2]{facet_grid}, "ggh4x" for \link[ggh4x]{facet_grid2}.}

\item{palette}{A vector of colors (bins, normal curve). See \code{vignette("palettes")}.}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{An interger. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{...}{Aditional arguments passed to the ggplot geom defined by \code{geom}.}

\item{independent}{For varying the scales of each cell. See \link[ggh4x]{facet_grid2}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
\code{ggvar_acf} plots the result of a \link[stats]{acf} call for every series, using ggplot and \link[ggplot2]{facet_wrap}. \code{ggvar_ccf} plots all the cross correlations (and similars) between the series, using ggplot in a matrix fashion with \link[ggplot2]{facet_grid}.
}
\examples{
ggvar_acf(EuStockMarkets)
ggvar_ccf(EuStockMarkets)
ggvar_acf(vars::VAR(EuStockMarkets))

}
