% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_dispersion.R
\name{ggvar_dispersion}
\alias{ggvar_dispersion}
\title{Plot VAR Residuals Dispersion}
\usage{
ggvar_dispersion(
  x,
  series = NULL,
  scales = "fixed",
  ncol = 1,
  alpha = 0.5,
  palette = c("black", "black")
)
}
\arguments{
\item{x}{A "varest" object to get residuals and fitted values from.}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{An interger. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{alpha}{A double. The alpha aesthetic for the points, passed to \link[ggplot2]{geom_point}.}

\item{palette}{A vector of colors (points, x-axis line). See \code{vignette("palettes")}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots a scatterplot of the residuals versus fitted values of a VAR model, using ggplot2.
}
\examples{
x <- vars::VAR(EuStockMarkets)
ggvar_dispersion(x)
}
