% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_distribution.R
\name{ggvar_distribution}
\alias{ggvar_distribution}
\title{Plot VAR Residuals Distribution}
\usage{
ggvar_distribution(
  x,
  series = NULL,
  plot_normal = TRUE,
  palette = c("grey", "black"),
  bins = 30,
  size = 1,
  ...
)
}
\arguments{
\item{x}{Either a "varest" object for plotting the residuals, or an dataset (object coercible to data.frame) with numeric variables.}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{plot_normal}{Logical, whether or not the normal curve should be plotted.}

\item{palette}{A vector of colors (bins, normal curve). See \code{vignette("palettes")}.}

\item{bins}{An integer. The number of histogram bins, passed to \link[ggplot2]{geom_histogram}.}

\item{size}{An integer. The size of the normal curve line, passed to \link[ggplot2]{geom_line}.}

\item{...}{Additional arguments passed to \link[ggplot2]{geom_histogram}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the histogram of the residuals of a VAR model, or of the variables in a dataset, overlapped with a normal curve.
}
\examples{
ggvar_distribution(vars::VAR(EuStockMarkets))

}
