% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_fevd.R
\name{ggvar_fevd}
\alias{ggvar_fevd}
\title{Plot for Forecast Error Variance Decomposition of a VAR}
\usage{
ggvar_fevd(
  x,
  n.ahead = NULL,
  series = NULL,
  geom = "bar",
  palette = NULL,
  scales = "fixed",
  ncol = 1,
  ...
)
}
\arguments{
\item{x}{A "varest" object to pass to \link[vars]{fevd}, or, directly, a "varfevd" object.}

\item{n.ahead}{An integer. The size of the forecast horizon, passed to \link[vars]{fevd}. Unused if `x` is "varfevd".}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{geom}{The ggplot geom used to create the plot, "bar" for \link[ggplot2]{geom_bar} (the default), "area" for \link[ggplot2]{geom_area}, or "line" for \link[ggplot2]{geom_line}.}

\item{palette}{A vector of colors for each variable. See \code{vignette("palettes")}.}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{An integer. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{...}{Additional arguments passed to the ggplot geom defined by \code{geom}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the result of a \link[vars]{fevd} call.
}
\examples{
ggvar_fevd(vars::VAR(freeny[-2]), n.ahead = 10)

}
