% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_fit.R
\name{ggvar_fit}
\alias{ggvar_fit}
\alias{ggvar_fit_colored}
\title{Plot VAR Fitted Values}
\usage{
ggvar_fit(
  x,
  compare = TRUE,
  series = NULL,
  index = 1:x$totobs,
  palette = c("black"),
  linetypes = c("solid", "dashed"),
  scales = "fixed",
  ncol = 1,
  ...
)

ggvar_fit_colored(
  x,
  compare = TRUE,
  series = NULL,
  index = 1:x$totobs,
  palette = NULL,
  linetypes = c("solid", "dashed"),
  ...
)
}
\arguments{
\item{x}{A "varest" object to get fitted values from.}

\item{compare}{Logical, should the true values be printed?}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{index}{A vector of labels to the x-axis, normally dates. Must have length equal to \code{x$obs}. Defaults to a numeric sequence.}

\item{palette}{A vector of colors. Just one for \code{ggvar_fit}, one for each variable for \code{ggvar_fit_colored}. See \code{vignette("palettes")}.}

\item{linetypes}{A vector of line types (fitted, original), passed to \link[ggplot2]{scale_linetype_manual}.}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{An integer. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{...}{Additional arguments passed to \link[ggplot2]{geom_line}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots fitted values of a VAR model, versus the actual values. \code{ggvar_fit} Plots each series in a facet. \code{ggvar_fit_colored} plots all in the same graph, each with a different color.
}
\examples{
x <- vars::VAR(freeny[-2])
ggvar_fit(x)
ggvar_fit_colored(x)

}
