% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_irf}
\alias{ggvar_irf}
\title{Plot Impulse Response Functions of a VAR}
\usage{
ggvar_irf(
  x,
  n.ahead = NULL,
  series_impulse = NULL,
  series_response = NULL,
  facet = "ggplot",
  ci = 0.95,
  ...,
  palette = c("black", "blue", "gray"),
  scales = "fixed",
  independent = "y"
)
}
\arguments{
\item{x}{A "varest" object to pass to \link[vars]{irf}, or, directly, a "varirf" object.}

\item{n.ahead}{An integer. The size of the forecast horizon, passed to \link[vars]{irf}. Unused if `x` is "varirf".}

\item{series_impulse}{A character vector with variables to consider for the impulses. Defaults to all (\code{NULL}).}

\item{series_response}{A character vector with variables to consider for the responses. Defaults to all (\code{NULL}).}

\item{facet}{The facet "engine" to be used. "ggplot2" for \link[ggplot2]{facet_grid}, "ggh4x" for \link[ggh4x]{facet_grid2}.}

\item{ci}{The level of confidence for the \link[vars]{irf}. Set to \code{FALSE} to omit.}

\item{...}{Further arguments passed to \link[vars]{irf}.}

\item{palette}{A vector of colors (bins, normal curve). See \code{vignette("palettes")}.}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{independent}{For varying the scales of each cell. See \link[ggh4x]{facet_grid2}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the result of a \link[vars]{irf} call.
}
\examples{
ggvar_irf(vars::VAR(EuStockMarkets), n.ahead = 10)

}
