% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_stability.R
\name{ggvar_stability}
\alias{ggvar_stability}
\title{Plot for Structural Stability of a VAR}
\usage{
ggvar_stability(
  x,
  series = NULL,
  ci = 0.95,
  ...,
  palette = c("black", "blue"),
  scales = "fixed",
  ncol = 1
)
}
\arguments{
\item{x}{A "varest" object to pass to \link[vars]{stability}, or, directly, a "varstabil" object.}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{ci}{The level of confidence for the \link[strucchange]{boundary}.}

\item{...}{Further arguments passed to \link[strucchange]{boundary}.}

\item{palette}{A vector of colors (line, conf. interval). See \code{vignette("palettes")}.}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{An integer. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the result of a \link[vars]{stability} call. Confidence intevals are calculated using \link[strucchange]{boundary}.
}
\examples{
ggvar_stability(vars::VAR(freeny[-2]))

}
