% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_values.R
\name{ggvar_values}
\alias{ggvar_values}
\alias{ggvar_values_colored}
\title{Plot Values of Dataset or VAR Residuals}
\usage{
ggvar_values(
  x,
  series = NULL,
  index = NULL,
  palette = c("black"),
  scales = "fixed",
  ncol = 1,
  ...
)

ggvar_values_colored(x, series = NULL, index = NULL, palette = NULL, ...)
}
\arguments{
\item{x}{Either a "varest" object for plotting the residuals, or an dataset (object coercible to data.frame) with numeric variables.}

\item{series}{A character vector with variables to consider. Defaults to all (\code{NULL}).}

\item{index}{A vector of labels to the x-axis, normally dates. Must have length equal to \code{x$obs} or \code{nrow(x)}. Defaults to a numeric sequence.}

\item{palette}{A vector of colors. Just one for \code{ggvar_values}, one for each variable for \code{ggvar_values_colored}. See \code{vignette("palettes")}.}

\item{scales}{"fixed" (the default), "free", "free_x" or "free_y". passed to \link[ggplot2]{facet_wrap}.}

\item{ncol}{An integer. The number of facet columns, passed to \link[ggplot2]{facet_wrap}.}

\item{...}{Additional arguments passed to \link[ggplot2]{geom_line}.}
}
\value{
An object of class \code{ggplot}.
}
\description{
Plots the historic values of variables in a dataset, or residuals of a VAR model. \code{ggvar_values} Plots each series in a facet. \code{ggvar_values_colored} plots all in the same graph, each with a different color.
}
\examples{
ggvar_values(EuStockMarkets)
ggvar_values_colored(EuStockMarkets)
ggvar_values(vars::VAR(EuStockMarkets))

}
