% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varutils-package.R
\docType{package}
\name{varutils-package}
\alias{varutils}
\alias{varutils-package}
\title{varutils: Utility Functions for VAR Analisys}
\description{
Functions for commonly needed ggplot graphs for Vector Auto-Regressive models, like IRF's, FEVD's, predict graphs, and VAR residual analisys. Soon to contain functions for common tests and statistics, and other VAR-like objects.
}
\keyword{internal}
